split tests.py into seperate test classes, and files
