+++ functional_tests/base.py
@@ -1,0 +1,29 @@
+import time
+
+from django.contrib.staticfiles.testing import StaticLiveServerTestCase
+from selenium import webdriver
+from selenium.common import WebDriverException
+
+MAX_WAIT = 10
+
+
+class FunctionalTest(StaticLiveServerTestCase):
+
+    def setUp(self) -> None:
+        self.browser = webdriver.Firefox()
+
+    def tearDown(self) -> None:
+        self.browser.quit()
+
+    def wait_for_row_in_list_table(self, row_text):
+        start_time = time.time()
+        while True:
+            try:
+                table = self.browser.find_element(by='id', value='id_list_table')
+                rows = table.find_elements(by='tag name', value='tr')
+                self.assertIn(row_text, [row.text for row in rows])
+                return
+            except (AssertionError, WebDriverException) as e:
+                if time.time() - start_time > MAX_WAIT:
+                    raise e
+                time.sleep(0.5)
